requirements.txt
fastapi
authx
pydantic
PyJWT
uvicorn
python-dotenv
